attached_assets/requirements_1771720416005.txt
pandas
plotly
kaleido
ipython
nbformat
ipykernel
snaptrade-python-sdk==10.34.10